requirements.txt
tensorflow >= 2.11
Pillow >= 9.4